ignore some more tags
These tags occur in the real dump but not in the
sample. The handling for unexpected tags
currently is to raise an exception, so these
needed to be ignored to allow dumps to be
processed to completion.

(No guarantees yet that all dumps will finish,
although the one from which the sample was
extracted has now been running for a fair while.)